[refactor] : bnv 패키지 이름 오류 수정
