+++ app/src/main/java/com.umc.yourweather/presentation/BnvHome.kt
@@ -1,10 +1,11 @@
-package com.umc.bottomnavi
+
+package com.umc.yourweather.presentation
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import com.umc.yourweather.R
 
 // TODO: Rename parameter arguments, choose names that match
@@ -31,8 +32,9 @@ class BnvHome : Fragment() {
     }
 
     override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?,
     ): View? {
         // Inflate the layout for this fragment
         return inflater.inflate(R.layout.fragment_bnv_home, container, false)
