+++ app/src/main/java/com.umc.yourweather/presentation/BnvReport.kt
@@ -1,10 +1,10 @@
-package com.umc.bottomnavi
+package com.umc.yourweather.presentation
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import com.umc.yourweather.R
 
 // TODO: Rename parameter arguments, choose names that match
@@ -31,8 +31,9 @@ class BnvReport : Fragment() {
     }
 
     override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?,
     ): View? {
         // Inflate the layout for this fragment
         return inflater.inflate(R.layout.fragment_bnv_report, container, false)
