+++ app/src/main/res/layout/activity_bnv.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BottomNavi">
+    tools:context=".presentation.BottomNavi">
 
     <FrameLayout
         android:id="@+id/fl_content"
